requirements.txt
flask==3.0.0
werkzeug==3.0.1
python-dotenv==1.0.0
openai==1.6.1
easyocr==1.7.1
python-dateutil==2.8.2
Pillow==10.1.0
PyMuPDF==1.23.8
torch==2.1.2
torchvision==0.16.2
